wip - implement react-datepicker
correct for browser time offset
