+++ package.json
@@ -48,7 +48,7 @@
     "prop-types": "^15.7.2",
     "query-string": "^6.13.7",
     "react": "^16.13.1",
-    "react-datepicker": "^7.6.0",
+    "react-date-picker": "^11.0.0",
     "react-dom": "^16.13.1",
     "react-draggable": "^4.4.3",
     "react-icons": "^4.11.0",
